requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
django-cors-headers
dotenv
langchain 
langchain-google-genai 
google-generativeai 
chromadb
langchain-community
djangorestframework-simplejwt

